requirements.txt
flask==2.0.1
python-dotenv==0.19.0
docx2pdf==0.1.8
Werkzeug==2.0.1
boto3==1.26.137
waitress==2.1.2
prometheus-client==0.19.0
gunicorn==21.2.0
python-docx==0.8.11
PyYAML==6.0.1
pywin32==306; sys_platform == 'win32'
